Add cli parsing to the bootstrap launcher and updater

+++ bootstrap/updater/src/main/scala/Updater.scala
@@ -6,6 +6,7 @@ import java.text.DecimalFormat
 import java.util.Date
 import java.util.zip.ZipFile
 
+import CLI._
 import me.tongfei.progressbar.{ProgressBar, ProgressBarBuilder, ProgressBarStyle}
 import org.fusesource.jansi.internal.CLibrary
 import org.json4s.jackson.JsonMethods.parse
@@ -22,7 +23,7 @@ object Updater {
   private val versionFileName = "version.txt"
   private val launcherJar = "bin/ChatOverflow-Launcher.jar"
   private val launcherMainClass = "Bootstrap"
-  private var classLoader = getLauncherLoader
+  private var classLoader: URLClassLoader = _
   private val ghBase = "https://api.github.com"
   private val acceptHeader = "Accept" -> "application/vnd.github.v3+json" // ensures that we always get the GitHub v3 API
   private implicit val jsonFormats: Formats = DefaultFormats
@@ -39,32 +40,50 @@ object Updater {
   def main(args: Array[String]): Unit = {
     println("Starting ChatOverflow Bootstrap Updater.")
 
-    println("Checking for updates...")
-    val update = Updater.searchForUpdates
-    if (update.isDefined) {
-      println("A new update is available!")
-      println(s"Current version: ${Updater.getCurrentVersion.get}")
-      println(s"Newest version: ${update.get.tag_name}")
+    val conf: Config = ArgsParser.parse(args, Config()) match {
+      case Some(value) => value
+      case None => System.exit(1); null
+    }
+    classLoader = getLauncherLoader(conf)
 
-      if (CLibrary.isatty(CLibrary.STDIN_FILENO) == 1) {
-        print(s"Do you want to download and install the update? [y/N] ")
-        val in = StdIn.readLine
-
-        if (in.toLowerCase == "y") {
-          val file = Updater.downloadUpdate(update.get)
-          if (file.isDefined) {
-            Updater.installUpdate(file.get)
-          }
-        }
-      } else {
-        println("Currently running in a non-interactive session. Please run in an interactive session to auto-update\n" +
-          s"or download and install manually from ${update.get.html_url}")
-      }
-    } else {
-      println("No new update is available.")
+    if (conf.help) {
+      // The user just wants to quickly get all available options, don't search for updates to not bother the user.
+      // Start the framework through the launcher and simply show the usage.
+      startLauncher(Array("--help"), conf)
+      return
     }
 
-    startLauncher(args)
+    if (!conf.ignoreUpdates) {
+      println("Checking for updates...")
+      val update = Updater.searchForUpdates
+      if (update.isDefined) {
+        println("A new update is available!")
+        println(s"Current version: ${Updater.getCurrentVersion.get}")
+        println(s"Newest version: ${update.get.tag_name}")
+
+        // Check if a tty is attached
+        if (CLibrary.isatty(CLibrary.STDIN_FILENO) == 1) {
+          print(s"Do you want to download and install the update? [y/N] ")
+          val in = StdIn.readLine
+
+          if (in.toLowerCase == "y") {
+            val file = Updater.downloadUpdate(update.get)
+            if (file.isDefined) {
+              Updater.installUpdate(file.get, conf)
+            }
+          }
+        } else {
+          println("Currently running in a non-interactive session. Please run in an interactive session to auto-update\n" +
+            s"or download and install manually from ${update.get.html_url}")
+        }
+      } else {
+        println("No new update is available.")
+      }
+    } else {
+      println("Skipping update check.")
+    }
+
+    startLauncher(args, conf)
   }
 
   /**
@@ -168,11 +187,12 @@ object Updater {
    * Installs a update from the passed zip file.
    *
    * @param zipFile the zip containing the update
+   * @param conf the cli config, used to extract the new files into the correct installation directory
    */
-  def installUpdate(zipFile: File): Unit = {
+  def installUpdate(zipFile: File, conf: Config): Unit = {
     println("Installing update...")
 
-    val binFiles = new File("bin").listFiles.filter(_.getName.endsWith(".jar"))
+    val binFiles = new File(s"${conf.directory}/bin").listFiles.filter(_.getName.endsWith(".jar"))
     binFiles.foreach(_.delete())
 
     classLoader.close() // release locks of jars on windows
@@ -183,7 +203,7 @@ object Updater {
     zip.entries().asScala
       .foreach(entry => {
         val is = zip.getInputStream(entry)
-        val out = new File(".", entry.getName)
+        val out = new File(conf.directory, entry.getName)
 
         if (out.isDirectory) {
           out.mkdirs()
@@ -195,7 +215,7 @@ object Updater {
               // Updater couldn't be updated, because Windows holds file locks on executing files including the updater.
               // Skip update of it, it shouldn't change anyway. We can update it on *nix system in the case we reeeeealy need to.
               // If it has changed and we can't auto-update, we do recommend the user to update the updater manually, but it is to the user to decide.
-              val currentIs = new BufferedInputStream(new FileInputStream("ChatOverflow.jar"))
+              val currentIs = new BufferedInputStream(new FileInputStream(s"${conf.directory}/ChatOverflow.jar"))
               val currentHash = Stream.continually(currentIs.read).takeWhile(_ != -1).map(_.toByte).hashCode()
               val newIs = new BufferedInputStream(zip.getInputStream(entry))
               val newHash = Stream.continually(newIs.read).takeWhile(_ != -1).map(_.toByte).hashCode()
@@ -213,12 +233,12 @@ object Updater {
       })
 
     // Re-set the executable flag for *nix systems
-    new File(".").listFiles()
+    new File(conf.directory).listFiles()
       .filter(f => f.isFile && f.getName.startsWith("ChatOverflow."))
       .foreach(_.setExecutable(true))
 
     // Reload all jar files, so that the launcher can be loaded
-    classLoader = getLauncherLoader
+    classLoader = getLauncherLoader(conf)
 
     println("Update installed.")
   }
@@ -240,30 +260,31 @@ object Updater {
    * Loads the class of the Launcher with the class loader and starts it.
    *
    * @param args the args to pass to the Bootstrap Launcher
+   * @param conf cli config, used to check the existence of the launcher jar
    */
-  def startLauncher(args: Array[String]): Unit = {
+  def startLauncher(args: Array[String], conf: Config): Unit = {
     try {
       val cls = classLoader.loadClass(launcherMainClass)
       val mainMethod = cls.getMethod("main", classOf[Array[String]])
       mainMethod.invoke(null, args)
     } catch {
       case _: ClassNotFoundException =>
-        if (!new File(launcherJar).exists)
+        if (!new File(s"${conf.directory}/$launcherJar").exists)
           println("Launcher jar is non existent. Seems like your installation is invalid.")
         else
           println(s"Main class of the launcher $launcherMainClass couldn't be found.")
       case _: NoSuchMethodException => println(s"Launcher jar is invalid: couldn't get main method.")
-      case e: Throwable => println(s"Launcher jar is invalid: $e")
+      case e: Throwable => println(s"Launcher jar is invalid: $e");
     }
   }
 
-  private def getLauncherLoader = new URLClassLoader(Array(
-    Paths.get(launcherJar).toUri.toURL
+  private def getLauncherLoader(conf: Config) = new URLClassLoader(Array(
+    Paths.get(s"${conf.directory}/$launcherJar").toUri.toURL
   ), getClass.getClassLoader.getParent)
 
   /**
    * Metadata about a release of ChatOverflow on GitHub.
-   * Note that the variables of this class doesn't represent all the metadata we get by the GitHub API,
+   * Note that the variables of this class don't represent all the metadata we get by the GitHub API,
    * you can add more metadata by adding a variable with the name that the metadata has in the json object.
    *
    * @param tag_name     the name of the git tag, which this release refers to
